+++ src/test/java/org/oogp/api/DataApi.java
@@ -31,7 +31,31 @@ public interface DataApi {
 
 	@Operation(
 		operationId = "getObjectAsString",
-		tags = { "DataApi" },
+		tags = { "Data" },
+		responses = {
+				@ApiResponse(
+					responseCode = "200",
+					content = {
+							@Content(mediaType = "application/json", schema = @Schema(implementation = String.class))
+					}),
+				@ApiResponse(
+					responseCode = "500",
+					content = {
+							@Content(mediaType = "application/json", schema = @Schema(implementation = ProblemDetail.class))
+					})
+		})
+	@GetMapping(
+		value = "object-string",
+		produces = { MediaType.APPLICATION_JSON_VALUE })
+	default ResponseEntity<Object> getObjectAsString() {
+		// this will generate schema with type: string because the @Schema implementation is defined as Object.class but
+		// spring-doc infers string because it cannot determine a structured schema
+		return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
+	}
+
+	@Operation(
+		operationId = "getStringAsObject",
+		tags = { "Data" },
 		responses = {
 				@ApiResponse(
 					responseCode = "200",
@@ -45,9 +69,9 @@ public interface DataApi {
 					})
 		})
 	@GetMapping(
-		value = "object-string",
+		value = "object-object",
 		produces = { MediaType.APPLICATION_JSON_VALUE })
-	default ResponseEntity<Object> getObjectAsString() {
+	default ResponseEntity<Object> getObjectAsObject() {
 		// this will generate schema with type: string because the @Schema implementation is defined as Object.class but
 		// spring-doc infers string because it cannot determine a structured schema
 		return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);